// src/app/tabs/shared/models/bill.model.ts
import { BillTransaction } from '../models/bill-transaction.model';
import { BillTariffDetail } from '../models/bill-tariff-detail.model';

export class Bill {
    id?: number;
    billNumber?: string;
    billDate?: Date;
    billDateLocal?: string;
    fromDate?: Date;
    fromDateLocal?: string;
    toDate?: Date;
    toDateLocal?: string;
    ownerId?: number;
    tenantId?: number;
    unitId?: number;
    unitNumber?: string;
    meterId?: number;
    previousReading?: number;
    presentReading?: number;
    previousReadingLocal?: string;
    presentReadingLocal?: string;
    formulaId?: number;
    formula?: string;
    consumptionRate?: number;
    remarks?: string;
    billPeriodId?: number;
    utilityTypeId?: number;
    billFormat?: string;
    termsConditionId?: number;
    termsCondition?: string;
    consumption?: number;
    consumptionUnit?: string;
    consumptionLocal: string;
    dueDate?: Date;
    billTypeId?: number;
    billType?: string;
    utilityType?: string;
    deviceName?: string;
    billAmount?: number;
    billAmountLocal?: string;
    paid?: number;
    toPay?: number;
    buildingName: string;
    ownerName?: string;
    sms?: boolean;
    email?: boolean;
    watsapp?: boolean;
    measuringUnit?: string;
    averageConsumption?: number;
    consumptionDifference?: number;
    status?: string;
    billTransactions?: BillTransaction[];
    billTariffDetails?: BillTariffDetail[];
    billGroupId?: number;
    isFailed?: boolean;
    consumptionDifferencePercentage?: number;
    consumptionDifferencePercentageLocal?: string;
    consumptionDifferenceLocal?: string;
    averageConsumptionLocal?: string;
    isPrimary?: boolean;
    isDifferentiateBill?: boolean;
    consumptionCharge?: string;
}